+++ neuralprophet/forecaster.py
@@ -17,7 +17,7 @@ from neuralprophet.plot_forecast_plotly import plot as plot_plotly
 from neuralprophet.plot_forecast_plotly import plot_components as plot_components_plotly
 from neuralprophet.plot_model_parameters_matplotlib import plot_parameters
 from neuralprophet.plot_model_parameters_plotly import plot_parameters as plot_parameters_plotly
-from neuralprophet.utils_plot import get_valid_configuration
+from neuralprophet.plot_utils import get_valid_configuration
 
 log = logging.getLogger("NP.forecaster")
